delete link

+++ main.py
@@ -3,6 +3,7 @@ import os
 from sys import argv
 
 FILE = "homepage.html"
+OUT_FILE = "out.html"
 
 def get_links():
     soup = get_soup()
@@ -24,12 +25,11 @@ def get_link(link_id):
     return None
 
 def add_link(link, name):
-    cout = len(get_links())
-    cout+=1
+    cout = len(get_links())+1
     file = open(FILE, "r")
     lines = file.readlines()
     file.close()
-    file_out = open("out_"+FILE, "a+");
+    file_out = open(OUT_FILE, "a+");
     for line in lines:
         if line != "</body>\n":
             file_out.write(line)
@@ -37,19 +37,37 @@ def add_link(link, name):
             file_out.write(f"<a index='{cout}' href='{link}'>{name}</a>\n</body>\n</html>\n")
             break
     file_out.close()
-    os.remove(FILE)
-    os.rename("out_"+FILE, FILE) 
+    remove_file()
 
 
-def delete_link():
-    #TODO
-    pass
+def remove_file():
+    if os.path.isfile(FILE):
+        os.remove(FILE)
+        os.rename(OUT_FILE, FILE) 
+    else:
+        return None
+
+def delete_link(index):
+    if get_link(index) != None:
+        file = open(FILE, "r")
+        lines = file.readlines()
+        file.close()
+        file_out = open(OUT_FILE, "a+")
+        for line in lines:
+            if line[1] == "a" and line[10] == index:
+                    continue
+            else:
+                file_out.write(line)
+        file_out.close()
+        remove_file()
 
 def main():
-    if (len(argv) == 1):
-        return 0
-    _, link, name = argv
-    add_link(link, name)
+    if (len(argv) == 2):
+        _, index = argv
+        delete_link(index)
+    elif (len(argv) == 3):
+        _, link, name = argv
+        add_link(link, name)
 
 if __name__ == "__main__":
     main()
